moved frtchAirlines() furtherup so landing page also have access to airlines
